Ceph Multibackend changes
1. Tempest has an option CONF.compute.compute_volume_common_az to make
sure that Volume and Instance are created in common AZ If Availability
zones are used. The paramater is respected when instance is created
because tempest.common.compute.create_test_server() is used but not
respected when volume is created for test purpose. Let's make sure
AZ is set when volume is created if parameter is set.

2. The plain Encryptor provider is not support on ceph backends, the
test test_encrypted_cinder_volumes_cryptsetup fails on:
cinder.exception.VolumeDriverException: Volume driver reported an error:
Provider plain not supported.
This comes from [1] which is called in [2]. The Plain provider will be deprecated
in future completely.

[1] https://opendev.org/openstack/cinder/src/branch/master/cinder/volume/volume_utils.py#L1195
[2] https://opendev.org/openstack/cinder/src/branch/master/cinder/volume/drivers/rbd.py#L877

Change-Id: Ib6290ed1ad5779e4e6fed449f3335dbdada42f0d