Temporarily delete two bad tests for consumers
These two tests are breaking the gate in barbican for a bugfix
patch. [1]

As currently written, the tests are checking the buggy behavior and they
break with the bugfix.

The main issue with the tests is that the different persona classes are
not overriding the cls.consumer_client object to match the persona under
test.  This results in the member client being used throughout, which
before the bufix was wrongly returning a 403.

These test are being re-written in a follow up patch. [2]

[1] https://review.opendev.org/c/openstack/barbican/+/816756/
[2] https://review.opendev.org/c/openstack/barbican-tempest-plugin/+/815071

Change-Id: I8698fc7a9ac849b8c24adfe824ca44dd3e42b999