Set a default region for clients
Clients initialized for tests.api.* tests do not have defined values
for region [1]. Because of that the clients connect to the first
available service endpoint that they find (see this line in
tempest [2]). This can cause issues when there are multiple endpoints
on a system for barbican (admin, internal, public).

This change sets a default value for key_manager.region config option
to 'regionOne' and ensures that clients defined on the side of
barbican-tempest-plugin are intialised with this value.

[1] https://opendev.org/openstack/barbican-tempest-plugin/src/branch/master/barbican_tempest_plugin/tests/api/base.py#L91
[2] https://opendev.org/openstack/tempest/src/branch/master/tempest/lib/auth.py#L586

Change-Id: Ic9ae00c663cca6b83dc961b984cf129d1c33afc4